add abstraction layer
